Fix classes after moving ValueValidator
