+++ MelonLoader/MelonLoader.csproj
@@ -103,6 +103,7 @@
     <Compile Include="Extensions\Shields\MonoMod\Hook_OnDetour.cs" />
     <Compile Include="Extensions\Shields\MonoMod\ILHook_OnDetour.cs" />
     <Compile Include="Extensions\UnityMagicMethods.cs" />
+    <Compile Include="Preferences\ValueValidator.cs" />
     <Compile Include="Utils\bHaptics\API.cs" />
     <Compile Include="Utils\bHaptics\NativeLibrary.cs" />
     <Compile Include="Libs\HarmonyX\Extras\DelegateTypeFactory.cs" />
