+++ MelonLoader/Preferences/ValueValidator.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace MelonLoader
+namespace MelonLoader.Preferences
 {
     public abstract class ValueValidator
     {
